# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UPER IDP API Documentation",
        "description": "API documentation for UPER IDP SSO System (CENTRAL)",
        "contact": {
            "email": "burhan.mafazi@universitaspertamina.ac.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Get list of roles",
                "operationId": "c1294bb85dfc016a8f9c0d3f0907293f",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search roles by name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of roles retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Roles retrieved successfully"
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "http://localhost:8000/api/v1/roles"
                                        },
                                        "method": {
                                            "type": "string",
                                            "example": "GET"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2023-06-01 10:00:00"
                                        },
                                        "total_data": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/RoleResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Create a new role",
                "operationId": "c7f58f49125e9b60da0f0db6890a4f22",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "display_name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "display_name": {
                                        "type": "string",
                                        "example": "Administrator"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "System administrator role"
                                    },
                                    "role_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "scope_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Role created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Role created successfully"
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "http://localhost:8000/api/v1/roles"
                                        },
                                        "method": {
                                            "type": "string",
                                            "example": "POST"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2023-06-01 10:00:00"
                                        },
                                        "total_data": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/RoleResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RoleResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "uuid": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "name": {
                        "type": "string",
                        "example": "admin"
                    },
                    "display_name": {
                        "type": "string",
                        "example": "Administrator"
                    },
                    "description": {
                        "type": "string",
                        "example": "System administrator role"
                    },
                    "role_type": {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "uuid": {
                                "type": "string",
                                "format": "uuid"
                            },
                            "code": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            },
                            "description": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "scope_type": {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "uuid": {
                                "type": "string",
                                "format": "uuid"
                            },
                            "code": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            },
                            "description": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with username and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}